arduino-python/requirements.txt
paho-mqtt==1.6.1
pyFirmata==1.1.0
pyserial==3.5
python-dotenv==0.21.0
